# backend/swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHome","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/health":{"get":{"operationId":"AppController_getHealth","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/people":{"get":{"operationId":"PeopleController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["People"]},"post":{"operationId":"PeopleController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePersonDto"}}}},"responses":{"201":{"description":""}},"tags":["People"]}},"/people/{id}":{"get":{"operationId":"PeopleController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["People"]},"patch":{"operationId":"PeopleController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePersonDto"}}}},"responses":{"200":{"description":""}},"tags":["People"]},"delete":{"operationId":"PeopleController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["People"]}},"/users":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Users"]},"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]},"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"]},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"description":"Credenciais para autenticação","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Token JWT e usuário","content":{"application/json":{"schema":{"example":{"access_token":"eyJhbGciOi...","user":{"id":1,"name":"Fulano","email":"usuario@wg.com","role":"ADMIN","active":true}}}}}},"401":{"description":"Credenciais inválidas."}},"summary":"Login de usuário","tags":["Autenticação"]}},"/auth/me":{"get":{"operationId":"AuthController_profile","parameters":[],"responses":{"200":{"description":"Retorna os dados do usuário autenticado."}},"security":[{"access-token":[]}],"summary":"Rota de teste protegida","tags":["Autenticação"]}}},"info":{"title":"WG Grappling API","description":"Sistema de gestão da academia WG Grappling","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"access-token":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreatePersonDto":{"type":"object","properties":{}},"UpdatePersonDto":{"type":"object","properties":{}},"CreateUserDto":{"type":"object","properties":{}},"UpdateUserDto":{"type":"object","properties":{}},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"usuario@wg.com","description":"E-mail do usuário"},"password":{"type":"string","example":"SenhaForte123","description":"Senha do usuário"}},"required":["email","password"]}}}}